requirements.txt
heic2png
Pillow
cx_Freeze